requirements.txt
opencv-contrib-python
numpy
imutils
urllib3
flask
opencv-python
python-dotenv
supabase
twilio